+++ ShibUI/ShibUI.lua
@@ -12,10 +12,9 @@ local sui = ShibUI
 sui.name        = "ShibUI"
 sui.menuName    = "ShibUI Settings"
 sui.displayName = "Shibui User Interface"
-sui.version     = "1.2.0" -- 1.2 updated ActionBar, 1.3 updated Compass
+sui.version     = "1.4.0" -- 1.2 updated ActionBar, 1.3 updated Compass, 1.4 updated GroupUnitFrame
 sui.author      = "Shownie & Ai"
 sui.description = "ShibUI is a minimalistic ESO addon that cleans up and modernizes the user interface."
--- end of global metadata
 
 --------------------------------------------------
 -- Main entry point for initializing ShibUI.
@@ -24,11 +23,10 @@ function sui.initialize()
     local initializers = {
         sui.initializeSettings,
         sui.initializeAttributeBar,
-        -- sui.initializeCompass,
         sui.initializeMiscellaneous,
         sui.initializeReloadUI,
         sui.initializeTargetBar,
-        sui.initializeUnitFrame,
+        -- sui.initializeUnitFrame,
     }
     for _, init in ipairs(initializers) do
         if type(init) == "function" then
@@ -54,6 +52,7 @@ end
     local initializers = {
         function() self.ActionBar:Initialize() end,
         function() self.Compass:Initialize() end,
+        function() self.UnitFrame:Initialize() end,
     }
     for _, initFunc in ipairs(initializers) do
         if type(initFunc) == "function" then
@@ -61,25 +60,16 @@ end
         end
     end
 end]]--
--- end of main entry point
 
 --------------------------------------------------
 -- Event handler for when the addon is loaded.
 --------------------------------------------------
-local function AddonLoaded(event, name)
-    if name ~= sui.name then return end
+local function AddonLoaded(event, addon)
+    if addon ~= sui.name then return end
     sui.initialize()
+    ShibUI.Compass:Initialize()
     ShibUI.ActionBar:Initialize()
     EVENT_MANAGER:UnregisterForEvent(sui.name, EVENT_ADD_ON_LOADED)
 end
 
 EVENT_MANAGER:RegisterForEvent(sui.name, EVENT_ADD_ON_LOADED, AddonLoaded)
-
---[[local function PlayerActivated(event, activated)
-    if not activated then return end
-    -- Initialize modules that need to wait until the player is fully loaded.
-    EVENT_MANAGER:UnregisterForEvent(sui.name, EVENT_PLAYER_ACTIVATED)
-end
-
-EVENT_MANAGER:RegisterForEvent(sui.name, EVENT_PLAYER_ACTIVATED, PlayerActivated)]]--
--- end of event handler